requirements.txt
altgraph==0.17.2
attrs==24.3.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
cssselect==1.2.0
cssutils==2.11.1
exceptiongroup==1.2.2
future==0.18.2
h11==0.14.0
idna==3.10
lxml==5.3.0
macholib==1.15.2
more-itertools==10.5.0
outcome==1.3.0.post0
premailer==3.10.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.27.1
six==1.15.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
yagmail==0.15.293
guara==0.0.10